Make Spark a provided dependency.
If you are running with a prebuilt distro (e.g., on the cloud, or on a cluster
running CDH), the spark-submit scripts will add the Spark JARs from the distro
to the classpath. As such, it is preferable to not build Spark JARs into the
user JAR you are providing.